🏗️ Move pages related to admin feature into new dedicated folder
