+++ src/features/admin/pages/Administration/index.jsx
@@ -1,0 +1,13 @@
+import { AuthenticatedLayout } from '../../../../layouts/AuthenticatedLayout'
+import { MembersList } from '../../members-list/components/MembersList'
+import { ScrollTopButton } from '../../../../components/Buttons/ScrollTopButton'
+import './style.scss'
+
+export function Administration() {
+    return (
+        <AuthenticatedLayout>
+            <MembersList />
+            <ScrollTopButton />
+        </AuthenticatedLayout>
+    )
+}
