--- src/pages/Administration/index.jsx
@@ -1,13 +1,0 @@
-import { AuthenticatedLayout } from '../../layouts/AuthenticatedLayout'
-import { MembersList } from '../../features/admin/members-list/components/MembersList'
-import { ScrollTopButton } from '../../components/Buttons/ScrollTopButton'
-import './style.scss'
-
-export function Administration() {
-    return (
-        <AuthenticatedLayout>
-            <MembersList />
-            <ScrollTopButton />
-        </AuthenticatedLayout>
-    )
-}
