+++ src/routes/Router.jsx
@@ -4,7 +4,7 @@ import { getUserOrganizationId } from '../features/user/store/userSelectors'
 import { UserProfile } from '../features/user/pages/UserProfile'
 import { Home } from '../pages/Home'
 import { OrganizationCreation } from '../features/organization/pages/OrganizationCreation'
-import { Administration } from '../pages/Administration'
+import { Administration } from '../features/admin/pages/Administration'
 import { ProfileSettings } from '../features/user/pages/ProfileSettings'
 import { Contact } from '../pages/Contact'
 import { SignUp } from '../features/user/pages/SignUp'
